Move icon out of main component
